Create concern fee_adaptations for common fee adaptor behaviour
