+++ app/services/ccr/fee/hardship_fee_adapter.rb
@@ -1,37 +1,8 @@
 module CCR
   module Fee
     class HardshipFeeAdapter < SimpleBillAdapter
+      include FeeAdaptations
       acts_as_simple_bill bill_type: 'AGFS_ADVANCE', bill_subtype: 'AGFS_HARDSHIP'
-    
-      def initialize(object = nil)
-        super(object) if object
-      end
-
-      def mappings
-        @mappings ||= fee_types.each_with_object({}) do |fee_type_unique_code, mappings|
-          mappings[fee_type_unique_code.to_sym] = { bill_type: bill_type, bill_subtype: bill_subtype }
-        end
-      end
-
-      def claimed?
-        filtered_fees.any? do |f|
-          f.amount&.positive? || f.quantity&.positive? || f.rate&.positive?
-        end
-      rescue NameError
-        raise ArgumentError, 'Instantiate with claim object to use this method'
-      end
-
-      private
-
-      def fee_types
-        %w[BABAF BADAF BADAH BADAJ BADAT BANOC BANDR BANPW BAPPE]
-      end
-
-      def filtered_fees
-        fees.select do |f|
-          fee_types.include?(f.fee_type.unique_code)
-        end
-      end
     end
   end
 end
